Remove possibly large number of DeferredRecipients attributes
RT 3.8.0 and 3.8.1 contained a bug, fixed in 9076df8b, which caused RT
to create a NULL DeferredRecipients Attribute for every outgoing email.
This not only bloats the database, but also causes problems with the RT
3.8.2 "content" upgrade step:

        my $attrs = RT::Attributes->new( RT->SystemUser );
        $attrs->UnLimit;
        my @dashboards = $attrs->Named('Dashboard');

This loads all Attribute records into memory, as Named does not perform
a Limit, but builds a hash of all results, and iterates over them.

Purge the NULL DeferredRecipients records, reducing database size and
preventing the difficulties that it causes with the above upgrade step.
This is done in both a 3.8.2 and 4.2.10 step -- the latter to reduce
database bloat for users who have already upgraded past 3.8.2 but ran on
3.8.0 or 3.8.1 for some time.

Fixes I#17357.