Sphinx 1.10 and above want an explicit binlog_path
Not explicitly setting binlog_path causes it to fall back to the
compiled-in default (often the current directory), which may cause
searchd to not start:

    http://sphinxsearch.com/docs/current.html#conf-binlog-path

Add an explicit binlog_path, accompanied by a comment that is is only
valid on 1.10 and above, as some distributions only ship 0.9.8.  This is
not done automatically by detecting the sphinx version, as the sphinx
binary may not be in PATH, or may be run from a remote server.

Tests are updated to adjust output from rt-setup-fulltext-index which
may be for the wrong version of sphinx.