Fix the Class::Accessor dependency
The dependency on Class::Accessor::Fast was first added in b5c65781, and
expanded in 1c3689a3.  However, rt-test-dependencies did not gain an
explicit Class::Accessor dependency until 9d552858, which was to support
370cd0d1 -- which was a use of Class::Accessor, not
Class::Accessor::Fast.

When the runtime-plugins branch was reverted (in 3b690954), 370cd0d1 was
reverted, but 9d552858 was not.  As such, since then RT has declared a
dependency on Class::Accessor, but makes use of Class::Accessor::Fast
instead.  Installs have only succeeded because some other dependency
pulls in Class::Accessor::Fast.

Switch to explicitly depending on Class::Accessor::Fast, as it is the
module that is in use by RT.